deploy_tmp/qb_app/requirements.txt
Flask
requests
pyodbc
python-dotenv
PyJWT
bcrypt
